+++ run_translation.py
@@ -264,7 +264,7 @@ class KNNArguments:
     dstore_dir: str = field(default="checkpoints")
     knn_sim_func: DIST.from_string = field(default=DIST.l2)
     lmbda: float = field(default=0.6)
-    k: float = field(default=16)
+    k: int = field(default=16)
     knn_temp: float = field(default=10)
     # Args for building the faiss index:
     build_index: bool = field(default=False)
